Move report_date_time() to external server as it isn't used by internal server anymore.
