+++ components/external_server/tests/routes/test_datamodel.py
@@ -5,10 +5,9 @@ from unittest.mock import Mock, patch
 
 import bottle
 
-from shared.utils.functions import md5_hash
-
-from routes import get_data_model
 from database.datamodels import default_source_parameters, default_subject_attributes
+from routes import get_data_model
+from utils.functions import md5_hash
 
 
 class DataModelTest(unittest.TestCase):
